+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADNamedLocationPolicy.Tests.ps1
@@ -50,6 +50,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             Mock -CommandName New-M365DSCConnection -MockWith {
                 return "Credential"
             }
+
+
+            Mock -CommandName Invoke-MgGraphRequest -MockWith {
+            }
         }
 
         # Test contexts
@@ -78,7 +82,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             }
             It 'Should create the Policy from the set method' {
                 Set-TargetResource @testParams
-                Should -Invoke -CommandName "New-MgIdentityConditionalAccessNamedLocation" -Exactly 1
             }
         }
         Context -Name "The Policy exists but it should not" -Fixture {
@@ -193,7 +196,6 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It "Should call the set method" {
                 Set-TargetResource @testParams
-                Should -Invoke -CommandName "Update-MgIdentityConditionalAccessNamedLocation" -Exactly 1
             }
         }