+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change log for Microsoft365DSC
 
+# 1.21.1117.1
+
+* EXOTransportRule
+  * Fixed issues with invalid State property and missing Enabled one;
+    ISSUE #1554;
+* IntuneDeviceCompliancePolicyAndroidDeviceOwner
+  * Initial release;
+* DEPENDENCIES
+  * Updated DSCParser to version 1.3.0.2;
+  * Updated MSCloudLoginAssistant to version 1.0.78;
+
 # 1.21.1110.1
 
 * MISC